Replace InputStream usages with File
Finding the class which will contain the bootstrap method might require consuming jar entries which would be obfuscated later, meaning a single pass over the bytes is not enough and random access is required. Note that this is only the case if one assumes that the class supposed to contain the bootstrap method might define a conflicting method, as otherwise class files can be obfuscated before the bootstrap method is actually added to its future owner.