Remove bootstrap method name conflict resolution
The current conflict resolution 'algorithm' would actually require three passes over the jar entries. A first one to resolve the correct name for the bootstrap method with eventual conflict resolution, a second pass for the actual obfuscation for which the chosen name of the bootstrap method is required, and a third pass to add the bootstrap method and library loading code to the owning class, as the library loading code may not be obfuscated itself. Instead, we now assume that the bootstrap method name is not taken and inform the user to choose a different one via --bootstrap-method-name on conflict.